requirements.txt
altair==5.2.0
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
bokeh==2.4.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
deprecation==2.1.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
extra-streamlit-components==0.1.60
fonttools==4.48.1
gitdb==4.0.11
GitPython==3.1.41
gotrue==2.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
Jinja2==3.1.3
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
pefile==2023.2.7
pillow==10.2.0
plotly==5.19.0
pluggy==1.4.0
postgrest==0.13.2
protobuf==4.25.2
pyarrow==14.0.2
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
PyJWT==2.8.0
pyparsing==3.1.1
pytest==8.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
realtime==1.0.2
referencing==0.32.1
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
st-login-form==0.2.1
st-supabase-connection==1.2.2
storage3==0.7.0
streamlit==1.32.2
streamlit-authenticator==0.1.5
StrEnum==0.4.15
supabase==2.3.3
supafunc==0.3.3
tenacity==8.2.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
websockets==11.0.3
xyzservices==2023.10.1
zipp==3.17.0
